+++ Classes/Controller/UploadController.php
@@ -89,6 +89,7 @@ final class UploadController extends ActionController
 
             // First add upload and commit to get uid
             $upload->setToken($this->tokenService->create($this->extensionKey . '_download_' . time()));
+            $upload->calculateAndSetValidityDate();
             $this->uploadRepository->add($upload);
             $this->uploadRepository->commit();